backend/app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-socketio[asgi]==5.11.4
tensorflow==2.20.0
opencv-python==4.10.0.84
numpy==1.26.4
